backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
black==24.1.1
CacheControl==0.14.2
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.1
exceptiongroup==1.2.2
fastapi==0.104.1
firebase-admin==6.2.0
flake8==7.0.0
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.67.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httplib2==0.22.0
httptools==0.6.4
idna==3.10
iniconfig==2.0.0
mccabe==0.7.0
msgpack==1.1.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.5.2
pydantic_core==2.16.2
pyflakes==3.2.0
PyJWT==2.10.1
pyparsing==3.2.1
pytest==8.0.1
python-dotenv==1.0.0
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
sniffio==1.3.1
starlette==0.36.3
tomli==2.2.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.2
